Right angle PCB holder which can be bolted or magnet fixed to index build plate
Allows PCBs to be repeatedly aligned to the same origin by bolting the holder to the build plate.  The opposite corner can be held using FDM0026 or this part using magnets.

Designed for 1.6mm (or thicker) PCB's which can be secured using the M3 holes and a M3+M5 washer.

Holds the face of the PCB at 10mm from build plate (1.6mm thick PCB)