Put fiducials in position file
As the global is set to use the "default" variant that we have set to be "kibom" type, the position files inherit the _mechanical filter which remove unuseful pads from the pick and place file.  Fiduals are hence not in the position file normally if the _mechanical filter is applied.

see the generic filter explanation;
https://github.com/INTI-CMNB/KiBot#supported-filters
and the supported variant default value of filter (`exclude_filter`) ;
https://github.com/INTI-CMNB/KiBot#supported-variants